Original Code
This is the original code I have. I still have more work to do on it such as allowing for the game to close without the game crashing because right now when I exit it the game just freezes and stops responding and I have to close the program from the taskbar. I also have a CSV file added because of future plans to implement a high score system. I have ideas for more features for the game as well.